Update and Correction of Japanese Translation (#98)
* Update Japanese Strings

I have updated, I need to be allowed to check the display on the app as I have not been able to confirm if there is a problem.

* Update values-ja

I have not yet confirmed how the display will look like, so I will fix it if there is a problem.

* Update strings.xml

* Update strings.xml

* Update strings.xml

* Update strings.xml

* Translation corrections and additions

Corrected missing parts of the translation and corrected errors in the Japanese translation.

Co-authored-by: Js0n <29531167+JasonKhew96@users.noreply.github.com>